+++ plugin-server/benchmarks/clickhouse/e2e.kafka.benchmark.ts
@@ -6,11 +6,10 @@ import { Hub, LogLevel, PluginsServerConfig, Queue } from '../../src/types'
 import { delay, UUIDT } from '../../src/utils/utils'
 import { makePiscina } from '../../src/worker/piscina'
 import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
-import { resetTestDatabaseClickhouse } from '../../tests/helpers/clickhouse'
+import { delayUntilEventIngested,resetTestDatabaseClickhouse } from '../../tests/helpers/clickhouse'
 import { resetKafka } from '../../tests/helpers/kafka'
 import { pluginConfig39 } from '../../tests/helpers/plugins'
 import { resetTestDatabase } from '../../tests/helpers/sql'
-import { delayUntilEventIngested } from '../../tests/shared/process-event'
 
 jest.setTimeout(600000) // 10min timeout
