+++ plugin-server/tests/clickhouse/postgres-parity.test.ts
@@ -1,5 +1,4 @@
 import { DateTime } from 'luxon'
-import { PoolClient } from 'pg'
 
 import { startPluginsServer } from '../../src/main/pluginsServer'
 import {
@@ -11,14 +10,13 @@ import {
     Team,
     TimestampFormat,
 } from '../../src/types'
-import { castTimestampOrNow, delay, UUIDT } from '../../src/utils/utils'
+import { castTimestampOrNow, UUIDT } from '../../src/utils/utils'
 import { makePiscina } from '../../src/worker/piscina'
 import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
-import { resetTestDatabaseClickhouse } from '../helpers/clickhouse'
+import { delayUntilEventIngested, resetTestDatabaseClickhouse } from '../helpers/clickhouse'
 import { resetKafka } from '../helpers/kafka'
 import { pluginConfig39 } from '../helpers/plugins'
 import { getFirstTeam, resetTestDatabase } from '../helpers/sql'
-import { delayUntilEventIngested } from '../shared/process-event'
 
 jest.mock('../../src/utils/status')
 jest.setTimeout(60000) // 60 sec timeout
